requirements.txt
numpy
scikit-image
matplotlib
imageio-ffmpeg
tensorboard
timm

chumpy
opencv-python

smplx
tqdm
yacs
contextlib2
joblib
configargparse
einops
